+++ lib/condition/constructor.go
@@ -63,7 +63,7 @@ var (
 	TypeCheckInterpolation = "check_interpolation"
 	TypeCount              = "count"
 	TypeJMESPath           = "jmespath"
-	TypeJSONSchema         = "jsonschema"
+	TypeJSONSchema         = "json_schema"
 	TypeNot                = "not"
 	TypeNumber             = "number"
 	TypeMetadata           = "metadata"
@@ -88,7 +88,7 @@ type Config struct {
 	CheckInterpolation CheckInterpolationConfig `json:"check_interpolation" yaml:"check_interpolation"`
 	Count              CountConfig              `json:"count" yaml:"count"`
 	JMESPath           JMESPathConfig           `json:"jmespath" yaml:"jmespath"`
-	JSONSchema         JSONSchemaConfig         `json:"jsonschema" yaml:"jsonschema"`
+	JSONSchema         JSONSchemaConfig         `json:"json_schema" yaml:"json_schema"`
 	Not                NotConfig                `json:"not" yaml:"not"`
 	Number             NumberConfig             `json:"number" yaml:"number"`
 	Metadata           MetadataConfig           `json:"metadata" yaml:"metadata"`